add new directories for sketches, extract mnm functionality
git-svn-id: https://mididuino.googlecode.com/svn/trunk@843 6f8c04f6-b8b6-11de-9846-1b058f2a8eb1

+++ hardware/libraries/GUI/RecordingEncoder.hh
@@ -27,7 +27,7 @@ public:
   void startRecording();
   void stopRecording();
   void clearRecording();
-  void playback(int pos);
+  void playback(uint8_t pos);
 
   virtual char *getName() {
     return realEnc->getName();
@@ -103,7 +103,7 @@ int RecordingEncoder<N>::update(encoder_t *enc) {
 }  
 
 template <int N>
-void RecordingEncoder<N>::playback(int pos) {
+void RecordingEncoder<N>::playback(uint8_t pos) {
   if (!playing)
     return;
